+++ crates/froglight-world/src/world/section/section.rs
@@ -15,13 +15,13 @@ use crate::world::{
 /// A [`Section`] is a `16x16x16 (X,Y,Z)` section of a
 /// [`Chunk`](crate::world::Chunk).
 ///
-/// A [`Section`] contains a counter for the number of non-air blocks in the
+/// Contains a counter for the number of non-air blocks in the
 /// section, and [`Containers`](ChunkDataContainer) for blocks and
 /// biomes.
 #[derive(Debug, Default, Clone, PartialEq, Eq, Hash, Reflect)]
 pub struct Section {
     /// The number of non-air blocks in the section.
-    pub block_count: usize,
+    pub block_count: u16,
     /// The block data stored in the section.
     pub blocks: ChunkDataContainer<BlockContainer>,
     /// The biome data stored in the section.
@@ -51,7 +51,7 @@ impl Section {
     /// Decodes a [`Section`] from a buffer.
     pub(crate) fn decode(buf: &mut Cursor<&[u8]>) -> Result<Self, ChunkDecodeError> {
         Ok(Self {
-            block_count: usize::from(u16::fg_read(buf)?),
+            block_count: u16::fg_read(buf)?,
             blocks: ChunkDataContainer::decode(buf)?,
             biomes: ChunkDataContainer::decode(buf)?,
         })
@@ -60,34 +60,31 @@ impl Section {
     /// Gets the block id at the given position in the section.
     #[must_use]
     #[inline]
-    pub fn get_blockid(&self, pos: SectionBlockPosition) -> usize { Self::get(&self.blocks, pos) }
+    pub fn get_blockid(&self, pos: SectionBlockPosition) -> u32 { Self::get(&self.blocks, pos) }
 
     /// Sets the block id at the given position in the section.
     ///
     /// Returns the previous block id at the position.
     #[inline]
-    pub fn set_blockid(&mut self, pos: SectionBlockPosition, value: usize) -> usize {
+    pub fn set_blockid(&mut self, pos: SectionBlockPosition, value: u32) -> u32 {
         Self::set(&mut self.blocks, pos, value)
     }
 
     /// Gets the biome id at the given position in the section.
     #[must_use]
     #[inline]
-    pub fn get_biomeid(&self, pos: SectionBlockPosition) -> usize { Self::get(&self.biomes, pos) }
+    pub fn get_biomeid(&self, pos: SectionBlockPosition) -> u32 { Self::get(&self.biomes, pos) }
 
     /// Sets the biome id at the given position in the section.
     ///
     /// Returns the previous biome id at the position.
     #[inline]
-    pub fn set_biomeid(&mut self, pos: SectionBlockPosition, value: usize) -> usize {
+    pub fn set_biomeid(&mut self, pos: SectionBlockPosition, value: u32) -> u32 {
         Self::set(&mut self.biomes, pos, value)
     }
 
     #[inline]
-    fn get<T: ContainerType>(
-        container: &ChunkDataContainer<T>,
-        pos: SectionBlockPosition,
-    ) -> usize {
+    fn get<T: ContainerType>(container: &ChunkDataContainer<T>, pos: SectionBlockPosition) -> u32 {
         container.get_data(&pos)
     }
 
@@ -96,8 +93,8 @@ impl Section {
     fn set<T: ContainerType>(
         container: &mut ChunkDataContainer<T>,
         pos: SectionBlockPosition,
-        value: usize,
-    ) -> usize {
+        value: u32,
+    ) -> u32 {
         container.set_data(&pos, value)
     }
